Fix build for macOS, add download and build instructions

+++ src/SimUtilities.cpp
@@ -2,7 +2,7 @@
 
 #include <QDateTime>
 
-#include "Assert.h"
+#include "AssertMacros.h"
 
 namespace mms {
 
@@ -20,20 +20,20 @@ QVector<TriangleGraphic> SimUtilities::polygonToTriangleGraphics(
     for (Triangle triangle : triangles) {
         TriangleGraphic graphic;
         graphic.p1 = {
-            triangle.p1.getX().getMeters(),
-            triangle.p1.getY().getMeters(),
+            static_cast<float>(triangle.p1.getX().getMeters()),
+            static_cast<float>(triangle.p1.getY().getMeters()),
             colorValues,
             alpha,
         };
         graphic.p2 = {
-            triangle.p2.getX().getMeters(),
-            triangle.p2.getY().getMeters(),
+            static_cast<float>(triangle.p2.getX().getMeters()),
+            static_cast<float>(triangle.p2.getY().getMeters()),
             colorValues,
             alpha,
         };
         graphic.p3 = {
-            triangle.p3.getX().getMeters(),
-            triangle.p3.getY().getMeters(),
+            static_cast<float>(triangle.p3.getX().getMeters()),
+            static_cast<float>(triangle.p3.getY().getMeters()),
             colorValues,
             alpha,
         };
